requirements.txt
dj-database-url==0.5.0
dj-email-url==1.0.6
Django==4.2.8
django-admin-search==0.3.15
django-allauth==0.58.2
django-cache-url==3.4.4
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-crispy-forms==2.0
django-datadownloader==1.2.1
django-extensions==3.2.3
django-filter==21.1
django-heroku==0.3.1
django-js-asset==2.2.0
django-mathfilters==1.0.0
django-phone-field==1.8.1
django-qrcode==0.3
django-rest-passwordreset==1.3.0
django-sendfile==0.3.11
django-taggit==5.0.1
django-waitress==0.1.0
djangorestframework==3.14.0
djangorestframework-jsonapi==4.3.0
dr-dump==1.1.5
drf-yasg==1.21.7
easy-thumbnails==2.8.5
environs==8.0.0
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.8.0
incremental==22.10.0
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
macholib @ file:///AppleInternal/Library/BuildRoots/9dd5efe2-7fad-11ee-b588-aa530c46a9ea/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
magic-filter==1.0.12
Markdown==3.3.6
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
oauthlib==3.2.2
packaging==23.1
Pillow==10.1.0
postgres==4.0
psycopg2==2.9.2
psycopg2-binary==2.9.7
psycopg2-pool==1.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pyOpenSSL==23.3.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytube==15.0.0
pytz==2021.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==23.1.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
soupsieve==2.5
SQLAlchemy==2.0.22
sqlparse==0.4.2
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2021.5
uritemplate==4.1.1
urllib3==2.0.6
waitress==1.4.4
Werkzeug==3.0.0
whitenoise==5.3.0
wikipedia==1.4.0
yarl==1.9.2
zipp==3.17.0
zope.interface==6.1
